Support CSS splitting for IE9.
Some tracks (like XO) that a large number of CSS rules get destroyed by IE9 because only half their CSS is ever applied. Old world atlas solved this by having two explicitly managed CSS files. New world electrode never accounted for this and the existing webpack setup mashes everything into one file.

Using `css-split-webpack-plugin` allows webpack to automatically split large CSS files into smaller ones that can be consumed be IE9 and lower. This process is automatic and transparent. The only caveat is that the original stylesheet is replaced with one that just imports the split stylesheets – this has to happen in electrode because it was decided that the app only has one CSS bundle file. :facepalm: If this was not the case the `imports` option could be turned off and `link` elements could be made for both split versions. And yes, source-maps are supported.

See:
 * https://github.com/metalabdesign/css-split-webpack-plugin
 * https://jira.walmart.com/browse/GPCC-6511
 * https://jira.walmart.com/browse/GPCC-6520