Implement basic PWA features (#40)
* Basic manifest.json and ServiceWorker support

* Remove extra SW registration script, change marker

Marker name changed so its consistent with the others

* Remove unused report.txt file

This was just a temp. thing so I could read through the output of a
console.log call that was super long

* Linting fixes

* Add 144x144 icon for web app install banner

* Inject meta tags from iconstats.json

iconstats.json is generated by favicon-webpack-plugin and includes a
manifest.json file. It contains all of the required icon assets, so all
we have to add is add the /js path so they resolve correctly.

Remove sw.js file

This was the result of an incorrect path for the build and should not be
tracked

* Remove generated stats files

* Remove unused service worker registration file

This will now be included in the client source and parses from the
webpack assets.

* Create service worker registration module

This file will be required using require.ensure so that it compiles to
its own file, for caching. It exports a function that must be called to
trigger the registration. The function accepts a callback that will be
called when the registration object is available.

Initially this used promises, but that caused the file size to go up a
lot since webpack was included polyfills for Promise + friends.

* Remove SW registration injection from webapp

This is handled by webpack now

* Return original paths from iconStats

We don't need to replace the path in the meta tags anymore since we're
specifying publicPath in the archetype now

* Remove build artifacts

* Add demo page for push notifications

* Use sw-config.js for all PWA configuration

* Remove unused register-sw.html file

* move js bundle to body

* Remove unused REGISTER_SW marker

* Return early if serviceWorker is not supported

* Add cURL command + more to push demo

The demo now has a curl command you can copy + pase into your shell to
send a push notification to the client! The demo will now also tell you
if SW is not supported, or if you're not subscribed to notifications
yet.

* Use arrow function with requre.ensure

Because we can!

* Rewrite service worker registration file

* Add real gcm_sender_id from firebase

This way we can send notifications from a remote endpoint

* Add meta tag for viewport (#2)

* add meta tag for viewport

* use decimal for scale

* Do not precache app icons

App icons should not be precached since they are client-specific, and no
one client will be requesting all of them. Precaching them all just
increases our total bytes transferred

* Use importScripts to extend service worker

This depends on work done in the archetype. It will parse the paths in
importScripts and create an entry point for each, then sw-precache will
import them directly.

This supports importing an arbitrary number of scripts from any location
in the client folder

* Fix curl command output in push demo

* Add icon and badge to push notifications

* Error handling for push notification demo